Fix VineTable dtype promotion in _enlarge_and_copy
np.zeros without dtype defaults to float64, causing all VineTable
arrays to promote from float16/bool to float64 on first resize.
This inflates visibility arrays by 8x (bool→float64).

+++ lib/models/trellis.py
@@ -686,7 +686,7 @@ class VineTable:
 
 
     def _enlarge_and_copy(self, x: np.ndarray, size: int) -> np.ndarray:
-        new_x = np.zeros([size, x.shape[1]]) 
+        new_x = np.zeros([size, x.shape[1]], dtype=x.dtype)
         new_x[:self.valid_cnt] = x[:self.valid_cnt]
         return new_x
